add heatmaplayer example to the website
