--- examples/website/heatmap-layer/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset='UTF-8' />
-    <title>deck.gl example</title>
-    <style type="text/css">
-      body {margin: 0; padding: 0; background: #fff;}
-    </style>
-  </head>
-  <body>
-    <script src='bundle.js'></script>
-  </body>
-</html>
