+++ examples/website/heatmap/webpack.config.js
@@ -1,42 +1,37 @@
 // NOTE: To use this example standalone (e.g. outside of deck.gl repo)
 // delete the local development overrides at the bottom of this file
 
-// avoid destructuring for older Node version support
-const resolve = require('path').resolve;
 const webpack = require('webpack');
 
 const CONFIG = {
+  mode: 'development',
+
   entry: {
-    app: resolve('./app.js')
+    app: './app.js'
   },
 
-  devtool: 'source-map',
+  output: {
+    library: 'App'
+  },
 
   module: {
     rules: [
       {
-        // Compile ES2015 using buble
+        // Transpile ES6 to ES5 with babel
+        // Remove if your app does not use JSX or you don't need to support old browsers
         test: /\.js$/,
-        loader: 'buble-loader',
-        include: [resolve('.')],
+        loader: 'babel-loader',
         exclude: [/node_modules/],
         options: {
-          objectAssign: 'Object.assign'
+          presets: ['@babel/preset-react']
         }
       }
     ]
-  },
-
-  resolve: {
-    alias: {
-      // From mapbox-gl-js README. Required for non-browserify bundlers (e.g. webpack):
-      'mapbox-gl$': resolve('./node_modules/mapbox-gl/dist/mapbox-gl.js')
-    }
   },
 
   // Optional: Enables reading mapbox token from environment variable
   plugins: [new webpack.EnvironmentPlugin(['MapboxAccessToken'])]
 };
 
-// This line enables bundling against src in this repo rather than installed deck.gl module
+// This line enables bundling against src in this repo rather than installed module
 module.exports = env => (env ? require('../../webpack.config.local')(CONFIG)(env) : CONFIG);
